reducing warnings

+++ trunk/firmware/controllers/math/avg_values.c
@@ -26,8 +26,8 @@ void avgAddValue(AvgTable *table, int rpm, float key, float value) {
 		return;
 	if (key >= MAX_KEY)
 		return;
-	int i = AVG_TAB_SIZE * rpm / MAX_RPM;
-	int j = AVG_TAB_SIZE * key / MAX_KEY;
+	int i = (int)(AVG_TAB_SIZE * rpm / MAX_RPM);
+	int j = (int)(AVG_TAB_SIZE * key / MAX_KEY);
 
 	table->values[i][j] += value;
 	table->counts[i][j]++;
@@ -45,8 +45,8 @@ float avgGetValue(AvgTable *table, int rpm, float key) {
 		return NAN;
 	if (key >= MAX_KEY)
 		return NAN;
-	int i = AVG_TAB_SIZE * rpm / MAX_RPM;
-	int j = AVG_TAB_SIZE * key / MAX_KEY;
+	int i = (int)(AVG_TAB_SIZE * rpm / MAX_RPM);
+	int j = (int)(AVG_TAB_SIZE * key / MAX_KEY);
 	return avgGetValueByIndexes(table, i, j);
 }
